+++ changelog.md
@@ -22,6 +22,8 @@ position in the `nextflow run` command inside the .nextflow.log file.
 - Added option to `dengue_typing` to retrive closest referece sequence and link it 
 with a secondary channel into `mafft`
 - New version of DEN-IM recipe
+- `integrity_coverage` now checks the integrity of the compressed read files with the 
+appropriate software. 
 
 
 ## 1.4.0